# [Perf] Use fastuuid for fast UUID generations - 2.1x Faster