--- a/.circleci/requirements.txt
+++ b/.circleci/requirements.txt
@@ -15,3 +15,4 @@ fastapi-sso==0.16.0
 uvloop==0.21.0
 mcp==1.10.1    # for MCP server  
 semantic_router==0.1.10 # for auto-routing with litellm
+fastuuid==0.12.0