## [Perf] Use fastuuid for fast UUID generations - 2.1x Faster 

**Performance analysis for model_response_creation operation**

| Metric | Before | After (Run 1) | After (Run 2) | After (Run 3) | Change (Run 1) | Change (Run 2) | Change (Run 3) |
|--------|--------|---------------|---------------|---------------|----------------|----------------|----------------|
| **model_response_creation Total Time** | 18.61s | 10.67s | 8.69s | 8.78s | **1.7x faster** | **2.1x faster** | **2.1x faster** |
| **model_response_creation Avg per Call** | 1.86ms | 1.07ms | 0.87ms | 0.88ms | **1.7x faster** | **2.1x faster** | **2.1x faster** |

<!-- e.g. "Implement user authentication feature" -->

## Relevant issues

<!-- e.g. "Fixes #000" -->

## Pre-Submission checklist

**Please complete all items before asking a LiteLLM maintainer to review your PR**

- [ ] I have Added testing in the [`tests/litellm/`](https://github.com/BerriAI/litellm/tree/main/tests/litellm) directory, **Adding at least 1 test is a hard requirement** - [see details](https://docs.litellm.ai/docs/extras/contributing_code)
- [ ] I have added a screenshot of my new test passing locally 
- [ ] My PR passes all unit tests on [`make test-unit`](https://docs.litellm.ai/docs/extras/contributing_code)
- [ ] My PR's scope is as isolated as possible, it only solves 1 specific problem


## Type

<!-- Select the type of Pull Request -->
<!-- Keep only the necessary ones -->

🆕 New Feature
🐛 Bug Fix
🧹 Refactoring
📖 Documentation
🚄 Infrastructure
✅ Test

## Changes