@CharlesB2 2025-09-05
⚠️ The fastuuid repository has been compromised, see https://github.com/fastuuid/fastuuid/issues/56. This should be reverted as soon as possible to avoid a potential supply chain attack. Reported to support@berri.ai as well.